게시물 수정시 권한체크

+++ src/main/java/com/exam/demo/controller/UsrArticleController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import javax.servlet.http.HttpSession;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -15,14 +16,9 @@ import com.exam.demo.vo.ResultData;
 
 @Controller
 public class UsrArticleController {
-//	@Autowired
+	@Autowired
 	private ArticleService articleService;
 
-	public UsrArticleController(ArticleService articleService) {
-		this.articleService = articleService;
-	}
-
-	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
 	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
@@ -36,14 +32,14 @@ public class UsrArticleController {
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 
 		if (Ut.empty(title)) {
-			return ResultData.from("F-1", "title을(를) 입력해주세요.");
+			return ResultData.from("F-1", "title을(를) 입력해주세요");
 		}
 		if (Ut.empty(body)) {
-			return ResultData.from("F-2", "body을(를) 입력해주세요.");
+			return ResultData.from("F-2", "body을(를) 입력해주세요");
 		}
 
 		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
@@ -51,7 +47,7 @@ public class UsrArticleController {
 
 		Article article = articleService.getArticle(id);
 
-		return ResultData.newData(writeArticleRd, article);
+		return ResultData.newData(writeArticleRd, "article", article);
 	}
 
 	@RequestMapping("/usr/article/getArticles")
@@ -59,7 +55,7 @@ public class UsrArticleController {
 	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
 
-		return ResultData.from("S-1", "게시물 리스트", articles);
+		return ResultData.from("S-1", "게시물 리스트", "articles", articles);
 	}
 
 	@RequestMapping("/usr/article/getArticle")
@@ -71,7 +67,7 @@ public class UsrArticleController {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물입니다.", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시물 입니다.", id), "articles", article);
 	}
 
 	@RequestMapping("/usr/article/doDelete")
@@ -87,27 +83,51 @@ public class UsrArticleController {
 		}
 
 		if (isLogined == false) {
-			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
 		}
 
 		Article article = articleService.getArticle(id);
-
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
+
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", "해당 게시물에 대한 권한이 없습니다");
+		}
+
 		articleService.deleteArticle(id);
-		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id), id);
+
+		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했습니다.", id), "id", id);
 	}
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+
 		Article article = articleService.getArticle(id);
+
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
-		articleService.modifyArticle(id, title, body);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물을 수정하였습니다.", id), id);
+		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
+
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
+		return articleService.modifyArticle(id, title, body);
 	}
+
 }
